% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_adfg.R
\name{theme_adfg}
\alias{theme_adfg}
\title{Format a Graph to ADF&G Publication Standards}
\usage{
theme_adfg(
  font_size = 18,
  font_family = "Times New Roman",
  rel_small = 0.86 * font_size/font_size,
  rel_tiny = 0.79 * font_size/font_size,
  rel_large = 1.15 * font_size/font_size,
  legend.position.set = c(0.8, 0.9),
  legend.justification = c("center"),
  strip.background = element_rect(fill = "grey80", color = NA),
  strip.placement = "outside",
  strip.text.y = element_text(angle = -90),
  box = "yes"
)
}
\description{
This function can be added to a ggplot to format the graph
to ADF&G formatting standards.
}
