% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_adfg.R
\name{theme_adfg}
\alias{theme_adfg}
\title{Format a Graph to ADF&G Publication Standards}
\usage{
theme_adfg(
  font_size = 18,
  font_family = "Times New Roman",
  font_size_title = 1.15,
  font_size_legend = 0.86,
  font_size_caption = 0.79,
  box = TRUE,
  base_line_size = font_size/22,
  ...
)
}
\arguments{
\item{font_size}{Size in points for font. Default is 18.}

\item{font_family}{Sets which font family you want to use. Default is
Times New Roman.}

\item{box}{TRUE/FALSE selection for whether to outline the plot in a box}

\item{...}{Other arguments that are standard to ggplot \code{theme()}}
}
\value{
The output returned will be either an object of \code{theme_adfg()},
}
\description{
\code{theme_adfg} creates a ggplot theme consistent for ADF&G publications.
}
\details{
Format a Graph to ADF&G Publication Standards

This function can be added to a ggplot to format the graph
to ADF&G formatting standards.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point() + theme_adfg()

}
